Moved exceptions to Transports; removed exception leaks
