+++ Libplanet.Net.Tests/Transports/DifferentAppProtocolVersionExceptionTest.cs
@@ -5,7 +5,7 @@ using System.Runtime.Serialization.Formatters.Binary;
 using Libplanet.Crypto;
 using Xunit;
 
-namespace Libplanet.Net.Tests
+namespace Libplanet.Net.Transports.Tests
 {
     public class DifferentAppProtocolVersionExceptionTest
     {
