+++ Libplanet.Net.Tests/Transports/InvalidCredentialExceptionTest.cs
@@ -4,7 +4,7 @@ using System.Runtime.Serialization.Formatters.Binary;
 using Libplanet.Crypto;
 using Xunit;
 
-namespace Libplanet.Net.Tests
+namespace Libplanet.Net.Transports.Tests
 {
     public class InvalidCredentialExceptionTest
     {
