+++ Libplanet.Net/Messages/MessageValidator.cs
@@ -3,6 +3,7 @@ using System.Collections.Immutable;
 using System.Globalization;
 using System.Linq;
 using Libplanet.Crypto;
+using Libplanet.Net.Transports;
 
 namespace Libplanet.Net.Messages
 {
