+++ Libplanet.Net/Messages/NetMQMessageCodec.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Bencodex;
 using Libplanet.Crypto;
+using Libplanet.Net.Transports;
 using NetMQ;
 
 namespace Libplanet.Net.Messages
