+++ Libplanet.Net/Transports/InvalidMessageTimestampException.cs
@@ -3,7 +3,7 @@ using System.Runtime.Serialization;
 using Libplanet.Net.Messages;
 using Libplanet.Serialization;
 
-namespace Libplanet.Net
+namespace Libplanet.Net.Transports
 {
     /// <summary>
     /// An exception that is thrown when the timestamp of an encoded <see cref="Message"/> is
